Refactor cmsApi testing
