+++ server/repositories/__tests__/cmsApi.spec.js
@@ -1,33 +1,56 @@
+/* eslint-disable class-methods-use-this */
 const nock = require('nock');
-const { DrupalJsonApiParams: Query } = require('drupal-jsonapi-params');
-const {
-  jsonApiResponse,
-  category,
-  tag,
-} = require('../../../test/resources/topics');
+const { jsonApiResponse } = require('../../../test/resources/jsonApi');
 const { JsonApiClient } = require('../../clients/jsonApiClient');
 const { CmsApi } = require('../cmsApi');
-const { TopicsQuery } = require('../cmsQueries/topicsQuery');
 
-const topicsQuery = new Query()
-  .addFields('taxonomy_term--tags', [
-    'drupal_internal__tid',
-    'name',
-    'description',
-  ])
-  .addFields('taxonomy_term--moj_categories', [
-    'name',
-    'description',
-    'field_legacy_landing_page',
-  ])
-  .addFilter(
-    'vid.meta.drupal_internal__target_id',
-    ['moj_categories', 'tags'],
-    'IN',
-  )
-  .addSort('name')
-  .addPageLimit(100)
-  .getQueryString();
+const path = `/jsonapi/test`;
+class TestQuery {
+  path() {
+    return path;
+  }
+
+  transform(item) {
+    const { name, description, fieldLegacyLandingPage } = item;
+    const id =
+      fieldLegacyLandingPage?.resourceIdObjMeta?.drupal_internal__target_id;
+    return {
+      id,
+      linkText: name,
+      description: description?.processed,
+      href: `/content/${id}`,
+    };
+  }
+}
+
+const testItem = ({ id, title, processed }) => ({
+  type: 'taxonomy_term--moj_test',
+  id: '8d9eaf09-a53e-42d9-a7be-2a2f04a0f315',
+  links: { self: { href: 'http://cms.prg/jsonapi/node/not-used' } },
+  attributes: {
+    name: title,
+    description: {
+      value: '<p>Not used</p>\r\n',
+      format: 'basic_html',
+      processed,
+    },
+  },
+  relationships: {
+    field_legacy_landing_page: {
+      data: {
+        type: 'node--landing_page',
+        id: 'e442003a-1f77-4d7f-ac90-b518878cfacd',
+        meta: {
+          drupal_internal__target_id: id,
+        },
+      },
+      links: {
+        self: { href: 'http://cms.prg/jsonapi/node/not-used' },
+        related: { href: 'http://cms.prg/jsonapi/node/not-used' },
+      },
+    },
+  },
+});
 
 describe('CmsApi', () => {
   let cmsApi;
@@ -47,12 +70,10 @@ describe('CmsApi', () => {
   });
 
   describe('get', () => {
-    const path = `/jsonapi/prison/berwyn/taxonomy_term?${topicsQuery}`;
-
     it('should handle returning no items', async () => {
       mockDrupal.get(path).reply(200, jsonApiResponse([]));
 
-      const response = await cmsApi.get(new TopicsQuery('berwyn'));
+      const response = await cmsApi.get(new TestQuery());
 
       expect(response).toStrictEqual([]);
     });
@@ -60,7 +81,7 @@ describe('CmsApi', () => {
     it('should propagate errors', () => {
       mockDrupal.get(path).reply(500, 'unexpected error');
 
-      return expect(cmsApi.get(new TopicsQuery('berwyn'))).rejects.toEqual(
+      return expect(cmsApi.get(new TestQuery())).rejects.toEqual(
         Error('Request failed with status code 500'),
       );
     });
@@ -68,7 +89,7 @@ describe('CmsApi', () => {
     it('propogates 404 as error', () => {
       mockDrupal.get(path).reply(404, 'unexpected error');
 
-      return expect(cmsApi.get(new TopicsQuery('berwyn'))).rejects.toEqual(
+      return expect(cmsApi.get(new TestQuery())).rejects.toEqual(
         Error('Request failed with status code 404'),
       );
     });
@@ -77,32 +98,21 @@ describe('CmsApi', () => {
       mockDrupal.get(path).reply(
         200,
         jsonApiResponse([
-          tag({
+          testItem({
             id: '1',
             title: 'One',
             processed: 'Desc 1',
           }),
-          category({
-            id: '2',
-            title: 'Two',
-            processed: 'Desc 2',
-          }),
         ]),
       );
-      const response = await cmsApi.get(new TopicsQuery('berwyn'));
+      const response = await cmsApi.get(new TestQuery());
 
       expect(response).toStrictEqual([
         {
           description: 'Desc 1',
-          href: '/tags/1',
+          href: '/content/1',
           id: '1',
           linkText: 'One',
-        },
-        {
-          description: 'Desc 2',
-          href: '/content/2',
-          id: '2',
-          linkText: 'Two',
         },
       ]);
     });
